refactor: apply review changes
