+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/processinstance/migration/MigrateProcessInstanceTest.java
@@ -20,8 +20,6 @@ import io.camunda.zeebe.protocol.record.value.BpmnElementType;
 import io.camunda.zeebe.test.util.BrokerClassRuleHelper;
 import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
-import java.time.Duration;
-import org.assertj.core.api.Assertions;
 import java.util.List;
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -205,620 +203,5 @@ public class MigrateProcessInstanceTest {
             "Expect that process definition path changed to contain new process definition key")
         .hasProcessDefinitionPath(List.of(v2ProcessDefinitionKey))
         .hasCallingElementPath(List.of());
-  }
-
-  @Test
-  public void shouldAdjustMessageCardinalityTrackingWhenMigratedForProcessInstance() {
-    final String processId = helper.getBpmnProcessId();
-    final String targetProcessId = helper.getBpmnProcessId() + "2";
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId)
-                    .startEvent("msg_start")
-                    .message("msg1")
-                    .serviceTask("A", t -> t.zeebeJobType("A"))
-                    .endEvent()
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(targetProcessId)
-                    .startEvent("msg_start")
-                    .message("msg2")
-                    .serviceTask("B", t -> t.zeebeJobType("B"))
-                    .endEvent()
-                    .done())
-            .deploy();
-
-    ENGINE.message().withName("msg1").withCorrelationKey("cardinality").publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start")
-            .withBpmnProcessId(processId)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    // when
-    final long targetProcessDefinitionKey =
-        extractProcessDefinitionKeyByProcessId(deployment, targetProcessId);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("A", "B")
-        .migrate();
-
-    // then
-    ENGINE.message().withName("msg1").withCorrelationKey("cardinality").publish();
-
-    Assertions.assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-                .withElementType(BpmnElementType.PROCESS)
-                .withBpmnProcessId(processId)
-                .withElementId(processId)
-                .skip(1) // skip the first activation
-                .exists())
-        .isTrue();
-  }
-
-  @Test
-  public void shouldAdjustMessageCardinalityTrackingWhenMigratedForTargetProcess() {
-    final String processId = helper.getBpmnProcessId();
-    final String targetProcessId = helper.getBpmnProcessId() + "2";
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId)
-                    .startEvent("msg_start")
-                    .message("msg1")
-                    .serviceTask("A", t -> t.zeebeJobType("A"))
-                    .endEvent()
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(targetProcessId)
-                    .startEvent("msg_start")
-                    .message("msg2")
-                    .serviceTask("B", t -> t.zeebeJobType("B"))
-                    .endEvent()
-                    .done())
-            .deploy();
-
-    ENGINE.message().withName("msg1").withCorrelationKey("cardinality").publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start")
-            .withBpmnProcessId(processId)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    // when
-    final long targetProcessDefinitionKey =
-        extractProcessDefinitionKeyByProcessId(deployment, targetProcessId);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("A", "B")
-        .migrate();
-
-    // then
-    final long messagePosition =
-        ENGINE
-            .message()
-            .withName("msg2")
-            .withCorrelationKey("cardinality")
-            .publish()
-            .getSourceRecordPosition();
-
-    // broadcast a record to have something to limit the stream on
-    final var limitPosition = ENGINE.signal().withSignalName("dummy").broadcast().getPosition();
-    Assertions.assertThat(
-            RecordingExporter.records()
-                .between(messagePosition, limitPosition)
-                .processInstanceRecords()
-                .withIntent(ProcessInstanceIntent.ELEMENT_ACTIVATED))
-        .extracting(r -> r.getValue().getElementId())
-        .describedAs("Expect that the target process is not activated")
-        .doesNotContain(targetProcessId);
-
-    // complete existing target instance
-    ENGINE.job().ofInstance(processInstanceKey).withType("A").complete();
-
-    // start new target instance
-    ENGINE.message().withName("msg2").withCorrelationKey("cardinality").publish();
-
-    Assertions.assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-                .withElementType(BpmnElementType.PROCESS)
-                .withBpmnProcessId(targetProcessId)
-                .exists())
-        .isTrue();
-  }
-
-  /**
-   * After migrating a process instance started by a message, we expect that a buffered message with
-   * the same correlation key cannot be correlated to the migrated instance until the process
-   * instance completes. We expect that the buffered message is automatically correlated once the
-   * migrated instance finishes.
-   */
-  @Test
-  public void shouldAdjustMessageCardinalityTrackingWhenMigratedForTargetProcessWithMessageTTL() {
-    final String processId = helper.getBpmnProcessId();
-    final String targetProcessId = helper.getBpmnProcessId() + "2";
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId)
-                    .startEvent("msg_start")
-                    .message("msg1")
-                    .serviceTask("A", t -> t.zeebeJobType("A"))
-                    .endEvent()
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(targetProcessId)
-                    .startEvent("msg_start")
-                    .message("msg2")
-                    .serviceTask("B", t -> t.zeebeJobType("B"))
-                    .endEvent()
-                    .done())
-            .deploy();
-
-    ENGINE.message().withName("msg1").withCorrelationKey("cardinality").publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start")
-            .withBpmnProcessId(processId)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    // when
-    final long targetProcessDefinitionKey =
-        extractProcessDefinitionKeyByProcessId(deployment, targetProcessId);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("A", "B")
-        .migrate();
-
-    // then
-    final long messagePosition =
-        ENGINE
-            .message()
-            .withName("msg2")
-            .withCorrelationKey("cardinality")
-            .withTimeToLive(Duration.ofMinutes(1)) // create message with TTL
-            .publish()
-            .getSourceRecordPosition();
-
-    // broadcast a record to have something to limit the stream on
-    final var limitPosition = ENGINE.signal().withSignalName("dummy").broadcast().getPosition();
-    Assertions.assertThat(
-            RecordingExporter.records()
-                .between(messagePosition, limitPosition)
-                .processInstanceRecords()
-                .withIntent(ProcessInstanceIntent.ELEMENT_ACTIVATED))
-        .extracting(r -> r.getValue().getElementId())
-        .describedAs("Expect that the target process is not activated")
-        .doesNotContain(targetProcessId);
-
-    // complete existing target instance
-    ENGINE.job().ofInstance(processInstanceKey).withType("A").complete();
-
-    Assertions.assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-                .withElementType(BpmnElementType.PROCESS)
-                .withBpmnProcessId(targetProcessId)
-                .withElementId(targetProcessId)
-                .exists())
-        .describedAs("Expect that the target process is activated for the message with TTL")
-        .isTrue();
-  }
-
-  @Test
-  public void shouldMigrateWithSameProcessIdWhenProcessInstanceIsCreatedWithMessageStartEvent() {
-    // given
-    final String processId1 = helper.getBpmnProcessId();
-    final String correlationKey = helper.getCorrelationValue();
-
-    ENGINE
-        .deployment()
-        .withXmlResource(
-            Bpmn.createExecutableProcess(processId1)
-                .startEvent("msg_start1")
-                .message("msg1")
-                .serviceTask("task1", t -> t.zeebeJobType("task"))
-                .done())
-        .deploy();
-    ENGINE.message().withName("msg1").withCorrelationKey(correlationKey).publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start1")
-            .withBpmnProcessId(processId1)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId1)
-                    .startEvent("msg_start")
-                    .message("msg2")
-                    .serviceTask("task2", t -> t.zeebeJobType("task"))
-                    .done())
-            .deploy();
-    final long targetProcessDefinitionKey =
-        MigrationTestUtil.extractProcessDefinitionKeyByProcessId(deployment, processId1);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("task1", "task2")
-        .migrate();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_MIGRATED)
-                .withProcessInstanceKey(processInstanceKey)
-                .withElementType(BpmnElementType.PROCESS)
-                .getFirst()
-                .getValue())
-        .describedAs("Expect that process definition key changed")
-        .hasProcessDefinitionKey(targetProcessDefinitionKey)
-        .describedAs("Expect that bpmn process id and element id is the same")
-        .hasBpmnProcessId(processId1)
-        .hasElementId(processId1)
-        .describedAs("Expect that version number is changed")
-        .hasVersion(2);
-  }
-
-  @Test
-  public void shouldMigrateWhenTargetHasNoMessageStartEvent() {
-    // given
-    final String processId1 = helper.getBpmnProcessId();
-    final String processId2 = helper.getBpmnProcessId() + "2";
-    final String correlationKey = helper.getCorrelationValue();
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId1)
-                    .startEvent("msg_start")
-                    .message("msg")
-                    .serviceTask("task1", t -> t.zeebeJobType("task"))
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId2)
-                    .startEvent("start")
-                    .serviceTask("task2", t -> t.zeebeJobType("task"))
-                    .done())
-            .deploy();
-    ENGINE.message().withName("msg").withCorrelationKey(correlationKey).publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start")
-            .withBpmnProcessId(processId1)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    final long targetProcessDefinitionKey =
-        MigrationTestUtil.extractProcessDefinitionKeyByProcessId(deployment, processId2);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("task1", "task2")
-        .migrate();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_MIGRATED)
-                .withProcessInstanceKey(processInstanceKey)
-                .withElementType(BpmnElementType.PROCESS)
-                .getFirst()
-                .getValue())
-        .describedAs("Expect that process definition key changed")
-        .hasProcessDefinitionKey(targetProcessDefinitionKey)
-        .describedAs("Expect that bpmn process id and element id changed")
-        .hasBpmnProcessId(processId2)
-        .hasElementId(processId2)
-        .describedAs("Expect that version number did not change")
-        .hasVersion(1);
-  }
-
-  @Test
-  public void shouldMigrateWhenSourceHasNoMessageStartEvent() {
-    // given
-    final String processId1 = helper.getBpmnProcessId();
-    final String processId2 = helper.getBpmnProcessId() + "2";
-    final String correlationKey = helper.getCorrelationValue();
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId1)
-                    .startEvent("start")
-                    .serviceTask("task1", t -> t.zeebeJobType("task"))
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId2)
-                    .startEvent("msg_start")
-                    .message("msg")
-                    .serviceTask("task2", t -> t.zeebeJobType("task"))
-                    .done())
-            .deploy();
-
-    final var processInstanceKey = ENGINE.processInstance().ofBpmnProcessId(processId1).create();
-
-    ENGINE.message().withName("msg").withCorrelationKey(correlationKey).publish();
-    RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-        .withElementType(BpmnElementType.START_EVENT)
-        .withElementId("msg_start")
-        .withBpmnProcessId(processId2)
-        .getFirst()
-        .getValue()
-        .getProcessInstanceKey();
-
-    final long targetProcessDefinitionKey =
-        MigrationTestUtil.extractProcessDefinitionKeyByProcessId(deployment, processId2);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("task1", "task2")
-        .migrate();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_MIGRATED)
-                .withProcessInstanceKey(processInstanceKey)
-                .withElementType(BpmnElementType.PROCESS)
-                .getFirst()
-                .getValue())
-        .describedAs("Expect that process definition key changed")
-        .hasProcessDefinitionKey(targetProcessDefinitionKey)
-        .describedAs("Expect that bpmn process id and element id changed")
-        .hasBpmnProcessId(processId2)
-        .hasElementId(processId2)
-        .describedAs("Expect that version number did not change")
-        .hasVersion(1);
-  }
-
-  @Test
-  public void shouldMigrateWhenSourceProcessInstanceIsCreatedWithoutCorrelationKey() {
-    // given
-    final String processId1 = helper.getBpmnProcessId();
-    final String processId2 = helper.getBpmnProcessId() + "2";
-    final String correlationKey = helper.getCorrelationValue();
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId1)
-                    .startEvent("msg_start")
-                    .message("msg1")
-                    .serviceTask("task1", t -> t.zeebeJobType("task"))
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId2)
-                    .startEvent("msg_start")
-                    .message("msg2")
-                    .serviceTask("task2", t -> t.zeebeJobType("task"))
-                    .done())
-            .deploy();
-    ENGINE.message().withName("msg1").withCorrelationKey("").publish();
-    ENGINE.message().withName("msg2").withCorrelationKey(correlationKey).publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start")
-            .withBpmnProcessId(processId1)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-        .withElementType(BpmnElementType.START_EVENT)
-        .withElementId("msg_start")
-        .withBpmnProcessId(processId2)
-        .getFirst()
-        .getValue()
-        .getProcessInstanceKey();
-
-    final long targetProcessDefinitionKey =
-        MigrationTestUtil.extractProcessDefinitionKeyByProcessId(deployment, processId2);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("task1", "task2")
-        .migrate();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_MIGRATED)
-                .withProcessInstanceKey(processInstanceKey)
-                .withElementType(BpmnElementType.PROCESS)
-                .getFirst()
-                .getValue())
-        .describedAs("Expect that process definition key changed")
-        .hasProcessDefinitionKey(targetProcessDefinitionKey)
-        .describedAs("Expect that bpmn process id and element id changed")
-        .hasBpmnProcessId(processId2)
-        .hasElementId(processId2)
-        .describedAs("Expect that version number did not change")
-        .hasVersion(1);
-  }
-
-  @Test
-  public void shouldMigrateWhenTargetProcessInstanceIsCreatedWithoutCorrelationKey() {
-    // given
-    final String processId1 = helper.getBpmnProcessId();
-    final String processId2 = helper.getBpmnProcessId() + "2";
-    final String correlationKey = helper.getCorrelationValue();
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId1)
-                    .startEvent("msg_start")
-                    .message("msg1")
-                    .serviceTask("task1", t -> t.zeebeJobType("task"))
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId2)
-                    .startEvent("msg_start")
-                    .message("msg2")
-                    .serviceTask("task2", t -> t.zeebeJobType("task"))
-                    .done())
-            .deploy();
-    ENGINE.message().withName("msg1").withCorrelationKey(correlationKey).publish();
-    ENGINE.message().withName("msg2").withCorrelationKey("").publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start")
-            .withBpmnProcessId(processId1)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-        .withElementType(BpmnElementType.START_EVENT)
-        .withElementId("msg_start")
-        .withBpmnProcessId(processId2)
-        .getFirst()
-        .getValue()
-        .getProcessInstanceKey();
-
-    final long targetProcessDefinitionKey =
-        MigrationTestUtil.extractProcessDefinitionKeyByProcessId(deployment, processId2);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("task1", "task2")
-        .migrate();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_MIGRATED)
-                .withProcessInstanceKey(processInstanceKey)
-                .withElementType(BpmnElementType.PROCESS)
-                .getFirst()
-                .getValue())
-        .describedAs("Expect that process definition key changed")
-        .hasProcessDefinitionKey(targetProcessDefinitionKey)
-        .describedAs("Expect that bpmn process id and element id changed")
-        .hasBpmnProcessId(processId2)
-        .hasElementId(processId2)
-        .describedAs("Expect that version number did not change")
-        .hasVersion(1);
-  }
-
-  @Test
-  public void shouldMigrateWhenTargetProcessInstanceIsCreatedWithDifferentCorrelationKey() {
-    // given
-    final String processId1 = helper.getBpmnProcessId();
-    final String processId2 = helper.getBpmnProcessId() + "2";
-    final String correlationKey1 = helper.getCorrelationValue() + "1";
-    final String correlationKey2 = helper.getCorrelationValue() + "2";
-
-    final var deployment =
-        ENGINE
-            .deployment()
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId1)
-                    .startEvent("msg_start")
-                    .message("msg1")
-                    .serviceTask("task1", t -> t.zeebeJobType("task"))
-                    .done())
-            .withXmlResource(
-                Bpmn.createExecutableProcess(processId2)
-                    .startEvent("msg_start")
-                    .message("msg2")
-                    .serviceTask("task2", t -> t.zeebeJobType("task"))
-                    .done())
-            .deploy();
-    ENGINE.message().withName("msg1").withCorrelationKey(correlationKey1).publish();
-    ENGINE.message().withName("msg2").withCorrelationKey(correlationKey2).publish();
-    final var processInstanceKey =
-        RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-            .withElementType(BpmnElementType.START_EVENT)
-            .withElementId("msg_start")
-            .withBpmnProcessId(processId1)
-            .getFirst()
-            .getValue()
-            .getProcessInstanceKey();
-
-    RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-        .withElementType(BpmnElementType.START_EVENT)
-        .withElementId("msg_start")
-        .withBpmnProcessId(processId2)
-        .getFirst()
-        .getValue()
-        .getProcessInstanceKey();
-
-    final long targetProcessDefinitionKey =
-        MigrationTestUtil.extractProcessDefinitionKeyByProcessId(deployment, processId2);
-
-    ENGINE
-        .processInstance()
-        .withInstanceKey(processInstanceKey)
-        .migration()
-        .withTargetProcessDefinitionKey(targetProcessDefinitionKey)
-        .addMappingInstruction("task1", "task2")
-        .migrate();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_MIGRATED)
-                .withProcessInstanceKey(processInstanceKey)
-                .withElementType(BpmnElementType.PROCESS)
-                .getFirst()
-                .getValue())
-        .describedAs("Expect that process definition key changed")
-        .hasProcessDefinitionKey(targetProcessDefinitionKey)
-        .describedAs("Expect that bpmn process id and element id changed")
-        .hasBpmnProcessId(processId2)
-        .hasElementId(processId2)
-        .describedAs("Expect that version number did not change")
-        .hasVersion(1);
   }
 }
